+++ src/pytask_parallel/execute.py
@@ -1,3 +1,4 @@
+"""Contains code relevant to the execution."""
 import sys
 import time