+++ CHANGES.rst
@@ -3,7 +3,8 @@ Changes
 
 This is a record of all past pytask-parallel releases and what went into them in reverse
 chronological order. Releases follow `semantic versioning <https://semver.org/>`_ and
-all releases are available on `Anaconda.org <https://anaconda.org/pytask/pytask-parallel>`_.
+all releases are available on `Anaconda.org
+<https://anaconda.org/pytask/pytask-parallel>`_.
 
 
 0.0.4 - 2020-10-30